requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
jinja2==3.1.4
python-dotenv==1.0.1
yt-dlp>=2026.3.17
pillow==11.0.0
aiosqlite==0.20.0
httpx==0.27.2
opencv-python>=4.13
numpy>=2.0
scipy>=1.14
imagehash>=4.3
